fix: changed types and server imlementation

+++ src/internal/models/types.go
@@ -1,6 +1,6 @@
 package models
 
 type Product struct {
-	Name string `json:"name" bson:"name"`
-	Price uint64 `json:"price" bson:"price"`
+	Name  string `json:"name" bson:"name"`
+	Price uint32 `json:"price" bson:"price"`
 }
